requirements.txt
browser-use-sdk
python-dotenv
pydantic>=2.0
pandas
